refactor: update directory structure and directory names
