+++ app/src/domain/ports/UserRepository.ts
@@ -1,4 +1,4 @@
-import User from '../entities/User'
+import User from '../model/User'
 
 export interface UserRepository {
   findByEmail(email: string): Promise<User | null>
