--- app/src/domain/ports/ValidationTokenPayload.ts
@@ -1,6 +1,0 @@
-export interface ValidationTokenPayload {
-  userId: string
-  email: string
-  iat: number
-  exp: number
-}
