+++ app/src/index.ts
@@ -1,11 +1,12 @@
 import express from 'express'
+import 'reflect-metadata'
 import { BcryptPasswordHasher } from './infrastructure/adapters/BCryptPasswordHasher'
 import { JwtTokenService } from './infrastructure/adapters/JwtTokenService'
 import { LoginUseCase } from './application/use-cases/LoginUseCase'
 import * as dotenv from 'dotenv'
 import { resolve } from 'path'
 import { AuthAdapter } from './infrastructure/adapters/AuthAdapter'
-import { MongoUserRepository } from './infrastructure/adapters/MongoUserRepository'
+import { MongoUserRepository } from './infrastructure/database/MongoUserRepository'
 import { RefreshTokenUseCase } from './application/use-cases/RefreshTokenUseCase'
 import { ValidationUseCase } from './application/use-cases/ValidationUseCase'
 import cookieParser from 'cookie-parser'
