+++ app/src/infrastructure/adapters/AuthAdapter.ts
@@ -1,6 +1,8 @@
 import { Request, Response } from 'express'
-import { LoginUseCase } from '../../application/use-cases/LoginUseCase'
-import { UserRepository } from '../../domain/ports/UserRepository'
+import axios from 'axios'
+import { LoginUseCasePort } from '../../domain/ports/LoginUseCasePort'
+import { RegistrationUseCasePort } from '../../domain/ports/RegistrationUseCasePort'
+import { ValidationUseCasePort } from '../../domain/ports/ValidationUseCasePort'
 import { RefreshTokenUseCasePort } from '../../domain/ports/RefreshTokenUseCasePort'
 import { ValidationUseCasePort } from '../../domain/ports/ValidationUseCasePort'
 import axios from 'axios'
