+++ app/src/infrastructure/database/MongoUserRepository.ts
@@ -1,5 +1,5 @@
 import { UserRepository } from '../../domain/ports/UserRepository'
-import User from '../../domain/entities/User'
+import User from '../../domain/model/User'
 import * as bcrypt from 'bcrypt'
 import mongoose, { Schema, Document, Model } from 'mongoose'
